test: pytest fixtures breakdown to support better structure
