+++ tests/cli/test_isolation.py
@@ -76,8 +76,6 @@ def test_file_modification_during_run(
     ]
     cmd = ["python", os.path.relpath(script, os.getcwd())]
 
-    previous = client.repo.head.commit
-
     with output.open("wb") as stdout:
         process = subprocess.Popen(prefix + cmd, stdin=subprocess.PIPE, stdout=stdout)
 
@@ -96,7 +94,3 @@ def test_file_modification_during_run(
 
     with output.open("r") as fp:
         assert "test" == fp.read().strip()
-
-    diff = previous.diff(client.repo.head.commit)
-    modifications = [modification for modification in diff if modification.change_type == "M"]
-    assert 0 == len(modifications)
