+++ tests/service/fixtures/service_jobs.py
@@ -1,0 +1,37 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright 2021 Swiss Data Science Center (SDSC)
+# A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+# Eidgenössische Technische Hochschule Zürich (ETHZ).
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Renku service fixtures for jobs testing."""
+import os
+
+import pytest
+
+
+@pytest.fixture
+@pytest.mark.usefixtures("svc_client", "mock_redis")
+def service_job(svc_client, mock_redis):
+    """Ensure correct environment during testing of service jobs."""
+    old_environ = dict(os.environ)
+
+    os.environ["RENKU_SVC_CLEANUP_TTL_FILES"] = "0"
+    os.environ["RENKU_SVC_CLEANUP_TTL_PROJECTS"] = "0"
+
+    try:
+        yield svc_client, mock_redis
+    finally:
+        os.environ.clear()
+        os.environ.update(old_environ)
