+++ tests/service/views/test_graph_views.py
@@ -21,14 +21,13 @@ import json
 import pytest
 from flaky import flaky
 
-from conftest import IT_REMOTE_REPO_URL
 from tests.service.views.test_dataset_views import assert_rpc_response
 
 
 @pytest.mark.service
 @pytest.mark.integration
 @flaky(max_runs=10, min_passes=1)
-def test_graph_build_view(svc_client_cache, authentication_headers_raw):
+def test_graph_build_view(svc_client_cache, authentication_headers_raw, it_remote_repo_url):
     """Create a new graph build job successfully."""
     svc_client, _, cache = svc_client_cache
 
@@ -38,7 +37,7 @@ def test_graph_build_view(svc_client_cache, authentication_headers_raw):
     assert "rq:job" not in cache_state
 
     payload = {
-        "git_url": IT_REMOTE_REPO_URL,
+        "git_url": it_remote_repo_url,
         "revision": "HEAD",
         "callback_url": "https://webhook.site",
     }
@@ -58,10 +57,10 @@ def test_graph_build_view(svc_client_cache, authentication_headers_raw):
 @pytest.mark.service
 @pytest.mark.integration
 @flaky(max_runs=1, min_passes=1)
-def test_graph_build_no_callback(svc_client_cache, authentication_headers_raw):
+def test_graph_build_no_callback(svc_client_cache, authentication_headers_raw, it_remote_repo_url):
     """Try to create a new graph build job."""
     svc_client, _, cache = svc_client_cache
-    payload = {"git_url": IT_REMOTE_REPO_URL, "revision": "HEAD"}
+    payload = {"git_url": it_remote_repo_url, "revision": "HEAD"}
 
     response = svc_client.post("/graph.build", data=json.dumps(payload), headers=authentication_headers_raw)
 
@@ -74,11 +73,11 @@ def test_graph_build_no_callback(svc_client_cache, authentication_headers_raw):
 @pytest.mark.service
 @pytest.mark.integration
 @flaky(max_runs=1, min_passes=1)
-def test_graph_build_no_revision(svc_client_cache, authentication_headers_raw):
+def test_graph_build_no_revision(svc_client_cache, authentication_headers_raw, it_remote_repo_url):
     """Create a new graph build job successfully."""
     svc_client, _, cache = svc_client_cache
 
-    payload = {"git_url": IT_REMOTE_REPO_URL, "callback_url": "http://localhost:8080"}
+    payload = {"git_url": it_remote_repo_url, "callback_url": "http://localhost:8080"}
 
     response = svc_client.post("/graph.build", data=json.dumps(payload), headers=authentication_headers_raw)
     assert response
